Add support for authenticating using Okta
This PR makes a few tweaks to add support for authenticating using Okta. It also includes the ability to disable OpenId authentication and legacy Stack Exchange authentication as well as adding the ability to require authentication (without requiring an allowlist of OpenIDs).

In order to test this you need to configure Okta with a client application and authorization server - I can provide those if needed.